pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "structkit"
version = "2.8.1"
description = "A structured data processing tool"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "httpdss"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
keywords = ["data", "structure", "processing"]

dependencies = [
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "openai>=1.0.0",
    "python-dotenv>=0.21.0",
    "jinja2>=3.1.0",
    "PyGithub>=1.58.0",
    "shtab>=1.6.0",
    "colorlog>=6.7.0",
    "pydantic-ai>=0.1.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
s3 = [
    "boto3>=1.26.0",
]
gcs = [
    "google-cloud-storage>=2.10.0",
    "google-api-core>=2.11.0",
]
cloud = [
    "structkit[s3]",
    "structkit[gcs]"
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.990",
]

[project.urls]
Homepage = "https://github.com/httpdss/structkit"
Repository = "https://github.com/httpdss/structkit.git"

[project.scripts]
structkit = "structkit.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]  # Look for packages in root directory
include = ["structkit*"]

[tool.setuptools.package-data]
structkit = ["contribs/*.yaml"]
